+++ CHANGELOG.rst
@@ -10,10 +10,20 @@ This change log uses principles from `keep a changelog <http://keepachangelog.co
 Added
 ^^^^^
 
+- Customizable landing page.
+- Logout button.
+- Search result pagniation.
+- Selection of entry number per page.
+- Version-dependent display of pagination.
+- Version-dependent display of README.yml content as clear text or as processed YAML/JSON.
+- A few simple unit tests.
 
 
 Changed
 ^^^^^^^
+
+- Updated from bootstrap-vue-3 to bootstap-vue-next
+- Adapted routes to remodeled dserver REST API.
 
 
 Deprecated
@@ -27,6 +37,7 @@ Removed
 Fixed
 ^^^^^
 
+- No search box on login screen.
 
 Security
 ^^^^^^^^